requirements.txt
# Note: Tesseract OCR must be installed separately on your system.
# For Windows, you can install it via Chocolatey:
#    choco install tesseract
# or download the installer from:
#    https://github.com/UB-Mannheim/tesseract/wiki
#
# Ensure that Tesseract is added to your system PATH so that pytesseract can use it.

annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
blis==1.2.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
cryptography==44.0.2
cymem==2.0.11
fastapi==0.115.11
Flask==3.1.0
gspread==6.2.0
google-auth==2.38.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mindee==4.18.0
murmurhash==1.0.12
numpy==2.2.3
packaging==24.2
pdfminer.six==20231228
pdfplumber==0.11.5
pdf2image==1.16.0
Pillow==10.0.0
preshed==3.0.9
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyMuPDF==1.22.5
pypdfium2==4.30.1
pytesseract==0.3.10
python-multipart==0.0.20
requests==2.32.3
rich==13.9.4
shellingham==1.5.4
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.0
streamlit
thinc==8.3.4
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
openai==0.27.8
